model/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
numpy==1.23.5
keras==2.12.0
mysql-connector-python==8.0.33
tensorflow==2.12.0
nltk==3.8.1
spacy==3.5.3
contractions==0.1.73
